More round and integers for python2

+++ test/integration/test_static_model.py
@@ -73,8 +73,8 @@ class BasicsTest(unittest.TestCase):
         assert np.all(d_c[:, cell_right] < 0)
 
         # Test fracture displacement
-        u_left = data['d_f'][:, :round(data['d_f'].shape[1]/2)]
-        u_right = data['d_f'][:, round(data['d_f'].shape[1]/2):]
+        u_left = data['d_f'][:, :int(round(data['d_f'].shape[1]/2))]
+        u_right = data['d_f'][:, int(round(data['d_f'].shape[1]/2)):]
         assert np.all(np.abs(u_left - u_right - 1) < 1e-10)
